test(agent): focus resume preparation regressions
